requirements.txt
# Base requirements
opencv-python>=4.5.0
pandas>=1.3.0

# Deep Learning frameworks
torch>=1.9.0
torchvision>=0.10.0

# YOLOv5 requirements
ultralytics>=8.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
PyYAML>=5.3.0
requests>=2.25.0
scipy>=1.7.0
tqdm>=4.62.0
Pillow>=8.0.0

# Database
mysql-connector-python>=8.0.0
fastapi~=0.115.5
uvicorn~=0.32.1
websockets~=14.1
python-dotenv~=1.0.1